pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tender-analyzer"
version = "1.0.0"
description = "Pipeline for tender analysis: email parsing, file downloading, text extraction, RAG classification, deep analysis"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "rank-bm25>=0.2.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "pandas>=1.5.0",
    "openpyxl>=3.1.0",
    "pdfplumber>=0.10.0",
    "python-docx>=0.8.11",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",
    "playwright>=1.40.0",
    "pywin32>=306; sys_platform == 'win32'",
    "rarfile>=4.1",
    "py7zr>=0.20.0",
    "nltk>=3.8",
    "scikit-learn>=1.3.0",
    "faiss-cpu>=1.7.4",
    "setuptools>=61.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0"
]

[project.scripts]
tender-pipeline = "pipeline.orchestrator:main"

[tool.setuptools.packages.find]
include = ["pipeline*"]
exclude = ["storage*", "tests*", "docs*"]